+++ lib/releaf/rspec.rb
@@ -1,2 +1,3 @@
+require 'releaf/rspec/matchers'
 require 'releaf/rspec/helpers'
 require 'releaf/rspec/features_matchers'