+++ releaf-content/app/builders/releaf/content/nodes/toolbox_builder.rb
@@ -14,19 +14,19 @@ module Releaf::Content::Nodes
     end
 
     def add_child_button
-      button(t('Add child'), "plus lg", class: "ajaxbox", href: url_for(action: "content_type_dialog", parent_id: resource.id))
+      button(t('Add child'), nil, class: "ajaxbox", href: url_for(action: "content_type_dialog", parent_id: resource.id))
     end
 
     def go_to_button
-      button(t('Go to'), "external-link lg", class: "ajaxbox", href: url_for(action: "go_to_dialog"))
+      button(t('Go to'), nil, class: "ajaxbox", href: url_for(action: "go_to_dialog"))
     end
 
     def copy_button
-      button(t('Copy'), "copy lg", class: "ajaxbox", href: url_for(action: "copy_dialog", id: resource.id))
+      button(t('Copy'), nil, class: "ajaxbox", href: url_for(action: "copy_dialog", id: resource.id))
     end
 
     def move_button
-      button(t('Move'), "arrows lg", class: "ajaxbox", href: url_for(action: "move_dialog", id: resource.id))
+      button(t('Move'), nil, class: "ajaxbox", href: url_for(action: "move_dialog", id: resource.id))
     end
 
   end