+++ releaf-core/lib/generators/dummy/templates/controllers/admin/books_controller.rb
@@ -1,7 +1,17 @@
 class Admin::BooksController < Releaf::BaseController
-  def resources
-    collection = super
-    collection = collection.where(active: true) if params[:only_active]
-    collection
+
+  def searchable_fields
+    [:title, :genre, author: [:name] ]
   end
+
+  def resources
+    relation = super
+
+    relation = relation.where(active: true) if params[:only_active]
+    relation = relation.where( 'published_at >= ?', params[:published_since]) if params[:published_since].present?
+    relation = relation.where( 'published_at <= ?', params[:published_up_to]) if params[:published_up_to].present?
+
+    relation
+  end
+
 end