+++ releaf-core/spec/features/ajaxbox_spec.rb
@@ -99,12 +99,13 @@ feature "Ajaxbox", js: true do
     expect(page).to_not have_css(".mfp-bg")
 
     find(".field[data-name='cover_image'] .value-preview img").click
-    full_image_url = find(".field[data-name='cover_image'] .value-preview a.ajaxbox")["href"]
-    image_url = full_image_url.match(/(\/media\/.*)/)[1] + "&ajax=1"
 
-    expect(page).to have_css(".ajaxbox-inner img[src='#{image_url}']")
+    image_url = find(".field[data-name='cover_image'] .value-preview a.ajaxbox")["href"] + "&ajax=1"
+    ajaxbox_image_selector = '.ajaxbox-inner img.mfp-img'
+    expect(find(ajaxbox_image_selector)['src']).to eq image_url
+
     find(".ajaxbox-inner button.close").click
-    expect(page).to_not have_css(".mfp-bg")
-    expect(page).to_not have_css(".ajaxbox-inner > img[src='#{image_url}']")
+    expect(page).to have_no_css(".mfp-bg")
+    expect(page).to have_no_css(ajaxbox_image_selector)
   end
 end