+++ releaf-core/spec/features/index_actions_spec.rb
@@ -3,7 +3,7 @@ feature "Base controller index", js: true do
   background do
     auth_as_user
     author = FactoryGirl.create(:author)
-    good_book = FactoryGirl.create(:book, title: "good book", author: author)
+    good_book = FactoryGirl.create(:book, title: "good book", author: author, published_at: Date.parse("2015-12-12"))
     FactoryGirl.create(:chapter, title: 'Scary night', text: 'Once upon a time...', book: good_book)
     FactoryGirl.create(:book, title: "bad book", author: author)
   end
@@ -22,6 +22,11 @@ feature "Base controller index", js: true do
     expect(page).to have_number_of_resources(0)
 
     uncheck "Only active"
+    expect(page).to have_number_of_resources(1)
+
+    fill_in "Published between", with: "2015-11-11"
+    click_button "Filter"
+
     expect(page).to have_number_of_resources(1)
   end
 
@@ -49,6 +54,7 @@ feature "Base controller index", js: true do
   scenario "keeps search parameters when navigating to edit and back" do
     visit admin_books_path(search: "good")
     click_link("good book")
+    wait_for_all_richtexts
     click_link("Back to list")
     expect(page).to have_number_of_resources(1)
   end